+++ xobjects/__init__.py
@@ -26,3 +26,5 @@ from .specialize_source import specialize_source
 from .typeutils import context_default, get_a_buffer
 
 from .dress import dress, JEncoder, DressedStruct
+
+from .linkedarray import BypassLinked
