+++ xobjects/context_cpu.py
@@ -17,6 +17,7 @@ from .context import (
     _concatenate_sources,
 )
 from .specialize_source import specialize_source
+from .linkedarray import BaseLinkedArray
 
 log = logging.getLogger(__name__)
 
@@ -72,44 +73,13 @@ def cdef_from_kernel(kernel, pyname=None):
     signature += ");"
     return signature
 
-class BypassLinked:
-    def __init__(self, container):
-        self.container = container
 
-    def __enter__(self):
-        self.container._flag_bypass_linked = True
-
-    def __exit__(self, *args, **kwargs):
-        del(self.container._flag_bypass_linked)
-
-class LinkedArrayCpu(np.ndarray):
-
-    container = None
-    container_setitem_name=None
-    mode = None
+class LinkedArrayCpu(BaseLinkedArray, np.ndarray): # order of bases defines which __setitem__ is used
 
     @classmethod
-    def from_array(cls, a, mode=None, container=None,
-                   container_setitem_name=None):
-        assert len(a.shape) == 1 # TODO: To be generalized
-        assert mode in (None, 'readonly', 'setitem_from_container')
-        self = cls(shape=a.shape, dtype=a.dtype, buffer=a.data, offset=0,
+    def _build_view(cls, a):
+        return cls(shape=a.shape, dtype=a.dtype, buffer=a.data, offset=0,
                    strides=a.strides, order='C')
-        self.mode=mode
-        self.container = container
-        self.container_setitem_name = container_setitem_name
-        return self
-
-    def __setitem__(self, indx, val):
-        print(f'{indx=} {val=}')
-        if (self.mode is None or
-            (hasattr(self.container, '_flag_bypass_linked')
-                         and self.container._flag_bypass_linked)):
-            super().__setitem__(indx, val)
-        elif self.mode == 'setitem_from_container':
-            getattr(self.container, self.container_setitem_name)(indx, val)
-        elif self.mode == 'readonly':
-            raise ValueError('This array is read only')
 
 class ContextCpu(XContext):
     """
